+++ examples/data-sources/linkridgecloud_services/data-source.tf
@@ -1,0 +1,7 @@
+provider "linkridgecloud" {
+  base_url = "https://dev.cloud.linkridge.net"
+}
+
+data "linkridgecloud_services" "qr" {
+  id = "qr-codes"
+}